+++ CHANGES.md
@@ -19,6 +19,8 @@ Unreleased
   scheduler.
 - Add `enqueue_at`, `enqueue_in`, and `cron_register` methods to the `@job`
   wrapper.
+- The worker runs jobs using `subprocess` and the `flask` CLI. This works on
+  Windows and macOS and avoids `fork` and `spawn`. {issue}`58`
 
 ## Version 0.3.3